+++ lib/src/license.dart
@@ -9,7 +9,7 @@ import 'dart:io';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 
-import 'license_detection/license_detector.dart' hide License;
+import 'license_detection/license_detector.dart' hide License, Range;
 import 'model.dart';
 
 const _licenseFileNames = ['LICENSE'];
@@ -55,7 +55,22 @@ Future<List<License>> detectLicenseInContent(
     return <License>[];
   }
 
-  return licenseResult.matches
-      .map((e) => License(path: relativePath, spdxIdentifier: e.identifier))
-      .toList();
+  return licenseResult.matches.map((e) {
+    return License(
+      path: relativePath,
+      spdxIdentifier: e.identifier,
+      range: Range(
+        start: Position(
+          offset: e.start.offset,
+          line: e.start.line,
+          column: e.start.column,
+        ),
+        end: Position(
+          offset: e.end.offset,
+          line: e.end.line,
+          column: e.end.column,
+        ),
+      ),
+    );
+  }).toList();
 }